move RoleTypeComponent to RoleTypes namespace
